+++ changelog.md
@@ -5,3 +5,9 @@ Every important update will be noted here
 ### Added
  - values() function
  - changelog
+
+## [1.0.39] - 2018-06-11
+### Added
+ - assassinate() function to remove values from objects and arrays
+ - omit() function to get a copy of an object minus a list of keys from an original object
+ - clone() function to deeply clone objects and arrays